Moved level extensions into level
